% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iviewMain.R
\name{appendImages}
\alias{appendImages}
\title{Append Images Horizontally or Vertically}
\usage{
appendImages(files, direction, file, run = TRUE,
  irfanViewExe = defaultIrfanViewExe())
}
\arguments{
\item{files}{paths to image files to be appended}

\item{direction}{1 = horizontally, 2 = vertically}

\item{file}{name of output file}

\item{run}{logical. If \code{TRUE} (default) the append action is performed,
otherwise the command line doing the append action is returned}

\item{irfanViewExe}{path to IrfanView Executable file}
}
\description{
Append Images Horizontally or Vertically
}
