% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iviewMain.R
\name{cropImageAndSetBorders}
\alias{cropImageAndSetBorders}
\title{Use IrfanView to Crop the Image and Set Borders}
\usage{
cropImageAndSetBorders(inputdir, outputdir,
  irfanViewExe = defaultIrfanViewExe(), crop = c(left = 0, top = 0,
  width = -1, height = -1), border = c(left = 0, right = 0, top = 0,
  bottom = 0), canvasColor = 255, useIni = TRUE, wait = FALSE,
  inputPattern = "*.gif")
}
\arguments{
\item{inputdir}{full input path including file name or pattern, e.g. "*.png",
"*.gif"}

\item{outputdir}{full output path including file name or pattern, e.g.
"*.pgm"}

\item{irfanViewExe}{path to IrfanView Executable File}

\item{crop}{vector of four integers (left, top, width, height): optional 
coordinates for cropping. If crop["left"] != -1 the image will be cropped, 
otherwise not.}

\item{border}{vector of four integers (left, right, top, bottom)}

\item{canvasColor}{passed to \code{kwb.iview:::writeIniFile}}

\item{useIni}{logical. If \code{TRUE} (default) an INI file is written
and IrfanView is called with the \code{/ini} option on the command line}

\item{wait}{passed to \code{shell}.}

\item{inputPattern}{file name pattern matching the files in \code{inputdir}
to be processed}
}
\description{
Use IrfanView to Crop the Image and Set Borders
}
