% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iviewMain.R
\name{iniContentCrop}
\alias{iniContentCrop}
\title{Content for INI File Specifying Crop Action}
\usage{
iniContentCrop(crop = c(left = 0, top = 0, width = -1, height = -1),
  border = c(left = 0, right = 0, top = 0, bottom = 0),
  canvasColor = 255)
}
\arguments{
\item{crop}{integer vector with elements named \code{left}, \code{top},
\code{width}, \code{height}}

\item{border}{integer vector with elements named \code{left}, \code{right},
\code{top}, \code{bottom}}

\item{canvasColor}{value used in IrvanView \code{[Effects]} option 
\code{CanvColor}}
}
\description{
Content for INI File Specifying Crop Action
}
