% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iviewMain.R
\name{iniFileAssignment}
\alias{iniFileAssignment}
\title{Assignement to be Used in INI File}
\usage{
iniFileAssignment(optname, optvalue)
}
\arguments{
\item{optname}{name of option}

\item{optvalue}{value of option}
}
\description{
Assignement to be Used in INI File
}
