% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iviewMain.R
\name{writeIniFile}
\alias{writeIniFile}
\title{Write INI File and Return Option String}
\usage{
writeIniFile(outputdir, crop, border, canvasColor, filename = "i_view32.ini")
}
\arguments{
\item{outputdir}{output directory}

\item{crop}{passed to \code{kwb.iview:::iniContentCrop}}

\item{border}{passed to \code{kwb.iview:::iniContentCrop}}

\item{canvasColor}{passed to \code{kwb.iview:::iniContentCrop}}

\item{filename}{name of INI file}
}
\value{
option string required to tell IrfanView to use the INI file
}
\description{
Write INI File and Return Option String
}
